scp should not delete source file after copy #1015

+++ core/src/main/java/com/dtolabs/rundeck/core/execution/impl/jsch/JschScpFileCopier.java
@@ -140,6 +140,7 @@ public class JschScpFileCopier extends BaseFileCopier implements FileCopier, Des
                                 script
                         );
 
+        final boolean removeFile = null == scriptfile;
 
 //        logger.debug("temp file for node " + node.getNodename() + ": " + temp.getAbsolutePath() + ",
 // datacontext: " + dataContext);
@@ -178,9 +179,11 @@ public class JschScpFileCopier extends BaseFileCopier implements FileCopier, Des
             context.getExecutionListener().log(0, errormsg);
             throw new FileCopierException("[jsch-scp] Failed copying the file: " + errormsg, failureReason, e);
         }
-        if (!localTempfile.delete()) {
-            context.getExecutionListener().log(Constants.WARN_LEVEL,
-                    "Unable to remove local temp file: " + localTempfile.getAbsolutePath());
+        if (removeFile){
+            if(!localTempfile.delete()){
+                context.getExecutionListener().log(Constants.WARN_LEVEL,
+                        "Unable to remove local temp file: " + localTempfile.getAbsolutePath());
+            }
         }
         return remotefile;
     }
